Move ContentRegisterHelpers
This makes them easier to use in the specs as well as the features
